Add conditional baselines for PHPStan
Based on the PHP version and installed package versions.

Added conditionally included baselines for PHP 5–7 and 8+, Composer 1 and 2, Dotenv 3–5.

This conditional functionality can be greatly reduced if we drop support for older versions of PHP, Composer, and other dependencies.

To resolve these conditional baselines, we have to replace the initial data of `Composer\InstalledVersions` which contains PHPStan's dependencies, since the context is its PHAR.

To accomplish this, we have to retrieve the contents of the project's installed dependencies from either:

1. `vendor/composer/installed.php` — Composer v2 PHP format.
2. `vendor/composer/installed.json` — Either Composer v1 or v2 JSON format.

The JSON format is tricky because it changes drastically between Composer v1 and v2 and the v2 PHP format. Both JSON formats must be remapped to the PHP format expected by `InstalledVersions`.

If the project's installed dependencies cannot be loaded, this file returns only the PHP baselines.

If the project's installed dependencies can be loaded, the extra baselines are resolved and the initial data of `InstalledVersions` is restored.